+++ frontend/src/components/Input.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import { StateInterface, useStore } from '../store';
+import { StateInterface, useFormStore } from '../store';
 
 const props = defineProps<{
     id: string
@@ -8,12 +8,12 @@ const props = defineProps<{
     model: keyof StateInterface
 }>()
 
-const st = useStore()
+const formStore = useFormStore()
 </script>
 
 <template>
     <div>
         <label v-if="label" :for="props.id">{{ props.label }}</label>
-        <input v-model="st[props.model]" :id="props.id" :type="props.type">
+        <input v-model="formStore[props.model]" :id="props.id" :type="props.type">
     </div>
 </template>
