+++ frontend/src/components/formElements/FormRow.vue
@@ -1,0 +1,14 @@
+<script setup lang="ts">
+const props = defineProps<{
+    twoCol?: boolean
+}>()
+
+const conditionalClass = { "two-col": props.twoCol }
+</script>
+
+
+<template>
+    <div class="row" :class=conditionalClass>
+        <slot></slot>
+    </div>
+</template>
